requirements.txt
charset-normalizer == 2.1.1
idna == 3.4
psycopg2-binary == 2.9.5
requests == 2.28.1
setuptools == 63.2.0
urllib3 == 1.26.13
vk-api == 11.9.9


